backend/requirements.txt
fastapi
uvicorn[standard]
groq
httpx
pydantic
python-dotenv
